Akk 15

+++ read_in_hdd.py
@@ -1,30 +1,26 @@
-import os
-from datetime import timedelta
+import argparse
 from glob import glob
-
-import re
 
 import shell
 from app import *
-
+from app.songs.functions import get_song_duration
 from app.songs.models import Song, Artist, Dance
 from app.users.models import User
-
-import argparse
-from mutagen.mp3 import MP3
-
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument("user", help="User name to use for filling the DB.")
     parser.add_argument("--base-path", help="Path to use for finding the files. If omitted, use the current pwd.",
                         default=os.getcwd())
+    parser.add_argument("--no-replace", help="Do not replace the old database file.", action="store_true")
 
     args = parser.parse_args()
 
-    os.unlink("app.db")
-
-    shell.create_db()
+    if not os.path.exists("app.db"):
+        shell.create_db()
+    elif not args.no_replace:
+        os.unlink("app.db")
+        shell.create_db()
 
     user = User.query.filter_by(name=args.user)
     if user.count() != 1:
@@ -55,20 +51,19 @@ if __name__ == '__main__':
                 print("File name {} in wrong format. Skipping.".format(file_name_with_this_dance))
                 continue
             else:
-                audio_file = MP3(file_name_with_this_dance)
-
                 dance_name, artist_name, title = dance_artist_title.split(" - ")
                 artist, artist_new_created = Artist.get_or_add_artist(artist_name)
 
                 new_song = Song(title, artist, dance, user)
                 new_song.path = file_name_with_this_dance.replace(base_path, "")[1:]
-                new_song.duration = timedelta(seconds=audio_file.info.length)
+                new_song.duration = get_song_duration(file_name_with_this_dance)
 
                 db.session.add(new_song)
-                db.session.commit()
 
                 added_counter += 1
 
             if added_counter % 10 == 0:
                 print("Added", added_counter, "song to the database")
+
+                db.session.commit()
